requirements.txt
blinker>=1.9.0
cryptography>=42.0.0
click>=8.3.0

Flask==3.1.3
Flask-Babel==4.0.0
flask-bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7

itsdangerous>=2.2.0
Jinja2>=3.1.6
MarkupSafe>=3.0.3
psycopg2-binary>=2.9.9
python-dotenv==1.2.1
Werkzeug>=3.1.6